feat(dependencies): add f2-client-ktor-http dependency to f2Client function
fix(imports): correct import paths for AuthRealmClientSecret and RealmId

refactor(build): remove redundant dependencies and add missing f2-client-domain

refactor(client): move ClientJvm class to organization-lib and delete from commons-domain

The changes add a new dependency to the f2Client function, ensuring the necessary modules are included. Import paths for AuthRealmClientSecret and RealmId are corrected to reflect their new locations. Redundant dependencies are removed from the build file, and missing dependencies are added to ensure proper functionality. The ClientJvm class is moved to the organization-lib module for better organization and modularity, and the redundant file in commons-domain is deleted.

refactor: update import paths for AuthRealm and RealmId

Update the import paths for `AuthRealm` and `RealmId` to reflect their new locations in `f2.client.domain`. This change ensures consistency and correctness in the codebase after the reorganization of the `f2.client` package. Additionally, update build.gradle.kts files to include the `f2Client` dependency for better modularization and dependency management.
