+++ im-commons/im-commons-domain/build.gradle.kts
@@ -6,11 +6,6 @@ plugins {
 
 dependencies {
     commonMainApi("io.komune.f2:f2-dsl-cqrs:${Versions.f2}")
+    commonMainApi("io.komune.f2:f2-client-domain:${Versions.f2}")
     Dependencies.Mpp.datetime(::commonMainApi)
-
-
-    Dependencies.Mpp.f2Client(::commonMainApi)
-    Dependencies.Mpp.ktor(::commonMainApi)
-    Dependencies.Js.ktor(::jsMainImplementation)
-    Dependencies.Jvm.ktor(::jvmMainImplementation)
 }
