+++ im-commons/im-commons-domain/src/commonMain/kotlin/io/komune/im/commons/model/AuthRealm.kt
@@ -1,6 +1,7 @@
 package io.komune.im.commons.model
 
-import f2.client.ktor.http.plugin.model.AuthRealm
+import f2.client.domain.AuthRealm
+
 
 data class AuthSubRealm(
     val master: AuthRealm,
