+++ im-f2/apikey/im-apikey-client/src/commonMain/kotlin/io/komune/im/f2/apikey/client/ApiKeyClient.kt
@@ -1,10 +1,10 @@
 package io.komune.im.f2.apikey.client
 
 import f2.client.F2Client
+import f2.client.domain.AuthRealm
 import f2.client.function
 import f2.client.ktor.F2ClientBuilder
 import f2.client.ktor.http.plugin.F2Auth
-import f2.client.ktor.http.plugin.model.AuthRealm
 import f2.dsl.fnc.F2SupplierSingle
 import f2.dsl.fnc.f2SupplierSingle
 import io.komune.im.apikey.domain.ApiKeyApi
