+++ im-f2/organization/im-organization-lib/build.gradle.kts
@@ -16,4 +16,7 @@ dependencies {
 
     implementation(project(Modules.Api.config))
     implementation(project(Modules.Infra.redis))
+
+    Dependencies.Mpp.f2Client(::implementation)
+    Dependencies.Jvm.ktor(::implementation)
 }
