+++ im-f2/organization/im-organization-lib/src/main/kotlin/io/komune/im/f2/organization/lib/service/InseeHttpClient.kt
@@ -1,7 +1,6 @@
 package io.komune.im.f2.organization.lib.service
 
 import io.komune.im.api.config.properties.InseeProperties
-import io.komune.im.commons.http.ClientJvm
 import io.komune.im.f2.organization.lib.model.insee.InseeResponse
 
 class InseeHttpClient(
