+++ im-f2/privilege/im-privilege-client/src/commonMain/kotlin/io/komune/im/f2/privilege/client/PrivilegeClient.kt
@@ -9,10 +9,10 @@ import io.komune.im.f2.privilege.domain.role.command.RoleDefineFunction
 import io.komune.im.f2.privilege.domain.role.query.RoleGetFunction
 import io.komune.im.f2.privilege.domain.role.query.RoleListFunction
 import f2.client.F2Client
+import f2.client.domain.AuthRealm
 import f2.client.function
 import f2.client.ktor.F2ClientBuilder
 import f2.client.ktor.http.plugin.F2Auth
-import f2.client.ktor.http.plugin.model.AuthRealm
 import f2.dsl.fnc.F2SupplierSingle
 import f2.dsl.fnc.f2SupplierSingle
 import io.ktor.client.plugins.HttpTimeout
