+++ im-f2/user/im-user-client/build.gradle.kts
@@ -6,4 +6,5 @@ plugins {
 
 dependencies {
     commonMainApi(project(Modules.F2.userDomain))
+    Dependencies.Mpp.f2Client(::commonMainApi)
 }
