+++ im-script/im-script-core/src/main/kotlin/io/komune/im/script/core/config/KeycloakClientProviderStub.kt
@@ -1,6 +1,6 @@
 package io.komune.im.script.core.config
 
-import f2.client.ktor.http.plugin.model.RealmId
+import f2.client.domain.RealmId
 import io.komune.im.api.config.bean.ImAuthenticationProvider
 import io.komune.im.commons.auth.currentAuth
 import io.komune.im.infra.keycloak.client.KeycloakClient
